JENA-44
QueryIterSort is now using a SortedDataBag to, eventually, spill on disk. The threshold to decide when to spill on disk is configured via ARQ symbol (i.e. spillOnDiskSortingThreshold) and by default the spilling is turned off. People can change the value of the threshold on a query level, dataset level or globally.

Thank you Sam for the initial patch and help in testing this, Stephen for the DataBags and Andy for all the feedback and help on this.

Similarly, we can commit and close JENA-45 now.

git-svn-id: https://svn.apache.org/repos/asf/incubator/jena/Jena2/ARQ/trunk@1167265 13f79535-47bb-0310-9956-ffa450edef68