JENA-195
HttpQuery and QueryEngineHTTP consistently throw a QueryExecException exception if a 'query' parameter is specified with SERVICE.
Added a couple of examples (i.e. ExampleDBPedia2|3) to show the two ways to specify a query parameter with SERVICE.

I hope everything is fine now. If not, let me know and I'll fix it directly in trunk.

git-svn-id: https://svn.apache.org/repos/asf/incubator/jena/Jena2/ARQ/trunk@1237753 13f79535-47bb-0310-9956-ffa450edef68